extensions/fun_stuff/requirements.txt
pillow
braillert
